Split up physics.engine into three smaller files wheeee
