+++ orbitx/physics/__init__.py
@@ -1,6 +1,6 @@
 """There's a lot of physics-related code, but all you really need are
 - the Physics Engine, physics.PhysicsEngine, and
 - miscellaneous calculation functions, physics.calc"""
-from . import engine
+from . import simulation
 
-PhysicsEngine = engine.PhysicsEngine
+PhysicsEngine = simulation.PhysicsEngine
